Refactor TypeScript definitions and update package.json paths
- Moved TypeScript definitions for the diff function from src/diff.d.ts to a new file diff.d.ts at the root level, improving organization.
- Updated package.json to reflect the new path for TypeScript definitions, ensuring compatibility and clarity in exports.
